Revert "all problems fixed"
