+++ src/App.jsx
@@ -3,13 +3,11 @@ import { Suspense, lazy } from 'react';
 import { Header } from './components/Header/Header';
 import { Loader } from './components/Loader/Loader';
 import RestrictedRoute from './components/RestrictedRoutes/RestrictedRoute';
-import { selectAuthIsLoading } from './redux/authSelectors';
 import { AppContainer } from './App.styled';
-import { useSelector } from 'react-redux';
 
 const WelcomePage = lazy(() => import('./pages/WelcomePage/WelcomePage'));
 const SignUpPage = lazy(() => import('./pages/SignUpPage/SignUpPage'));
-const SignInPage = lazy(() => import('./pages/SignInPage/SignInPage'));
+const SigninPage = lazy(() => import('./pages/SignInPage/SigninPage'));
 const HomePage = lazy(() => import('./pages/HomePage/HomePage'));
 
 
@@ -27,7 +25,7 @@ const appRoutes = [
     path: '/signin',
     element: (
       <RestrictedRoute>
-        <SignInPage />
+        <SigninPage />
       </RestrictedRoute>
     ),
   },
@@ -38,21 +36,18 @@ const appRoutes = [
 ];
 
 const App = () => {
-  const isLoading = useSelector(selectAuthIsLoading);
-
   return (
     <>
       <Header />
-      {isLoading ? (<Loader />) :
-        (<AppContainer>
-          <Suspense fallback={<Loader />}>
-            <Routes>
-              {appRoutes.map(({ path, element }) => (
-                <Route key={path} path={path} element={element} />
-              ))}
-            </Routes>
-          </Suspense>
-        </AppContainer>)}
+      <AppContainer>
+        <Suspense fallback={<Loader />}>
+          <Routes>
+            {appRoutes.map(({ path, element }) => (
+              <Route key={path} path={path} element={element} />
+            ))}
+          </Routes>
+        </Suspense>
+      </AppContainer>
     </>
   );
 };
