+++ src/components/Loader/Loader.styled.js
@@ -2,5 +2,5 @@ import styled from '@emotion/styled';
     
 export const StyledLoader = styled.div`
  margin-top: 200px;
- text-align: center; 
+ text-align: center;
 `
